backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
channels==4.2.2
channels_redis==4.2.1
Django==5.2
djangorestframework==3.16.0
msgpack==1.1.0
psycopg2-binary==2.9.10
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.13.2
